UnrealMathTest Failing on Startup due to change in QuatToRot in UnrealMath.
In changelist 3873301 I changed how Quat's converted to Rotator's to make them more predictable for content creation workflows with the sequencer in particularly with the singularities at Pitch at 90,-90.

The UnrealMathTest directly copies code for that test and was not updated and so would fail with VS2017.  In VS2015 it seems that ClampAxis would convert the 360.0 degrees for the test but due to slight precision issues that don't show up in 2017. Correcting the test is the correct fix here.
Tested fix on both VS2015 and VS2017.

#jira UE-56194

#rb ben.zeigler, max.chen
#lockdown nick.penwarden

[CL 3950302 by Mike Zyracki in Main branch]