When calculating vertical extents find feature points where slopes are zero and check them in addition to the keys if the curve is cubic. Curves now fit correclty vertically.
Also changed fudge to 5% from 10% to match up old editor. Tighter fits seems better.

AndrewR I left the adding of the snap vallue factor to the max calculation though I'm not 100% happy with it since it's not discoverable that's the reason you can't zoom in on your curve.  I would vote to remove it since you can always zoom out and then snap.

#jira UE-58571

#rb Andrew.Rodham

[CL 4049429 by Mike Zyracki in Dev-Sequencer branch]