#ROBOMERGE-AUTHOR: mike.zyracki Fix for performance issues with motion trails.
Found that the main issue was the GetRefTM call.  Refactored the code so that we only call that once per tick, instead of up to  4 times per tick. Also we cache the calculation of the key positions, so we don't need to calculate them again when calculating keys.  Finally the GetRefTM call was recursive which was a decent hint, so we cache out it's parent's first and just get the transform's linearly.

Also we don't select parents or children to show their trails also, this seemed buggy and possibly very very slow.

#jira UE-58712
#rb max.chen

#ROBOMERGE-SOURCE: CL 4100305 in //UE4/Release-4.20/...
#ROBOMERGE-BOT: RELEASE (Release-4.20 -> Release-Staging-4.20)

[CL 4100309 by mike zyracki in Staging-4.20 branch]