Duplicate 4098091
On Mac we can avoid continually reallocating the depth-fallback texture as we only need the largest size to hang around and we can use DontCare for Load & Store. On iOS due to driver bugs we need to match the size and either Clear or Load - but I think Clear should be faster. Should make FortGPUTestBed faster on the CPU.

#rb none
#jira NA

[CL 4101500 by Mark Satterthwaite in 4.20 branch]