Fixed crash when running in standalone editor build using sub-instances that copy properties
Revamped change as the last one failed Fast Path automated tests. This change limits the property re-find only to cases where we are using sub-instances.
DestProperty serialization was always a bad idea, this is another artifact of the problem where the property persists but the class it belongs to gets trashed. Re-mapping the property by name to the correct outer fixes this for a minor startup cost.

#jira UE-59510 - Crash in Standalone When Sub Anim Instance Uses Exposed Variable
#rb Jurre.deBaare

[CL 4109096 by Thomas Sarkanen in 4.20 branch]