#ROBOMERGE-AUTHOR: matt.collins #jira UERNDR-406 Bringing CL 4104051 from Dev-Rendering (//UE4/Dev-Rendering) to Release-4.20 (//UE4/Release-4.20)
---
Some fixes for semaphore handling.

Moved the completion handler to EndFrame (out of FlushFreeList())

FlushFreeList() attached a handler to the command buffer that signaled the semaphore. This can be called through RHIFlushResources which will cause a mismatched signal()

Added BeginFrame/EndFrame to the loop in AddModalWindows so the renderer gets the expected frame delimiters. Removed call to EndFrame from EndDrawingViewport.

#rb mark.satterthwaite, matt.kuhlenschmidt
---

#ROBOMERGE-SOURCE: CL 4111106 in //UE4/Release-4.20/...
#ROBOMERGE-BOT: RELEASE (Release-4.20 -> Release-Staging-4.20)

[CL 4112521 by matt collins in Staging-4.20 branch]