UGS: Added a dialog which shows all the required platform SDKs for a branch, linked from the status panel in UGS.
The llist is configured via the UGS config file submitted to Engine/Programs/UnrealGameSync/UnrealGameSync.ini (and may be overridden by the project config file if necessary):

    [Default]
    ; Set this to a network share which contains the SDK installers for your site
    SdkInstallerDir=

    ; All the required SDKs for the current version of the engine
    +SdkInfo=(Category="Android", Description="NDK r21", Browse="$(SdkInstallerDir)\\Android")
    +SdkInfo=(Category="Windows", Description="Visual Studio 2017")
    +SdkInfo=(Category="Windows", Description="Visual C++ Toolchain 14.13.26128")
    +SdkInfo=(Category="Windows", Description="Windows SDK 10.0.16299.0")

Similar entries for console platforms are added in console subdirectories. Each entry may contain an Install="Foo.exe" and/or Browse="C:\Foo" style attribute, specifying the path to an installer to run or directory to open in explorer respectively.

The SdkInstallerDir setting is used as a base directory for the default installers, seen above for Android. Licensees may override this with a network path specific to the site that UGS is being deployed to (either in this file, in a project specific config file, or in a Engine/Programs/UnrealGameSync/NotForLicensees/UnrealGameSync.ini file).

#rb none

[CL 4184311 by Ben Marsh in Dev-Core branch]