FNiagaraCompileRequestData::GetReferencedObjects Making sure that the cloned graph makes it onto the list of saved objects.
UNiagaraScript::BeginCacheForCookedPlatformData
Commandlets like DerivedDataCacheCommandlet call BeginCacheForCookedPlatformData directly on objects. This may mean that we have not properly gotten the HLSL script generated by the time that we get here. This does the awkward work of waiting on the parent system to finish generating the HLSL before we can begin compiling it for the GPU.

UNiagaraEditorModule::OnPreGarbageCollection
For commandlets like GenerateDistillFileSetsCommandlet, they just load the package and do some hierarchy navigation within it tracking sub-assets, then they garbage collect. Since nothing is holding onto the system at the root level, it will be summarily killed and any of references will also be killed. To thwart this for now, we are forcing the compilations to complete BEFORE garbage collection kicks in. To do otherwise for now has too many loose ends (a system may be left around after the level has been unloaded, leaving behind weird external references, etc). This should be revisited when more time is available (i.e. not days before a release is due to go out).

#jira UE-61248
#rb frank.fella
#tests auto tests pass

[CL 4186429 by Shaun Kime in 4.20 branch]