Duplicate 4190391:
Fix a long-standing error in MetalRHI that allowed for thread-unsafe operations to occur when binding an SRV or UAV - it would call GetLinearTexture which would potentially allocate. This should not have been happening - the allocation is meant to occur in the SRV/UAV constructor. This CL makes that separation explicit and will simply assert if the expected Linear Texture is not present.

#rb none
#jira N/A

[CL 4190604 by Mark Satterthwaite in 4.20 branch]