Duplicate 4190391:
Fix a long-standing error in MetalRHI that allowed for thread-unsafe operations to occur when binding an SRV or UAV - it would call GetLinearTexture which would potentially allocate. This should not have been happening - the allocation is meant to occur in the SRV/UAV constructor. This CL makes that separation explicit and will simply assert if the expected Linear Texture is not present.

#rb none
#jira N/A

#ROBOMERGE-SOURCE: CL 4190604 in //UE4/Release-4.20/...
#ROBOMERGE-BOT: RELEASE (Release-4.20 -> Release-Staging-4.20)
#ROBOMERGE-AUTHOR: mark.satterthwaite

[CL 4195547 by mark satterthwaite in Staging-4.20 branch]