UE-61351 Error: Failed to send -1 bytes of data to RemoteConnection spammed in the editor log when closing RemoteSession on device during connection
#jira UE-61351
#4.21
#rb Jack.Porter

Not an real issue. The message is issued whenever the host (or the client) tries to send messages through an inactive channel. There is a "/ping" message sent back and forth between the host and the client, with high frequency (hence the spam) - if this message is no longer received, the host (or the client) decides to terminate the connection.

"Downgrading" the message from Error to Log, to avoid spamming the log on release builds.

[CL 4339523 by Sorin Gradinaru in Dev-Mobile branch]