UHT header resolve optimization
GUnrealSourceFilesMap used to contain absolute paths. Map lookup was done via filename, which always failed. It then traversed the entire map and tried to match via module relative paths, which also always failed. It finally prepended a slash to the filename and did another full scan and checked if entries ended with the slash + filename. Changed so that the map contains filename only to start with.

#rb stefan.boberg, steve.robb

[CL 4394856 by Johan Torp in Dev-Core branch]