Sequencer: Fixed compilation of tracks with empty space where a compile range overlapped both empty space and an existing compiled segment.
The issue is that previously we only reported empty space when compiling a specific _time_, which meant that the inference that empty space must exist where no overlapping track segments existed was correct. However, with the shift to compiling the entire evaluated range, this inference is incorrect and a more robust approach is required. We now iterate the unique segments in the track for the whole compilation range in exactly the same way that we iterate sub sections, adding empty space only where no compiled segments exist (or could be compiled) in the track.

Also added a new test (Empty Space On The Fly) to verify this failure case, which failed under the old code.

#jira UE-65672
#rb Max.Chen

#ROBOMERGE-SOURCE: CL 4503437 in //UE4/Release-4.21/...
#ROBOMERGE-BOT: RELEASE (Release-4.21 -> Release-Staging-4.21)

[CL 4503442 by andrew rodham in Staging-4.21 branch]