UGS: Allow scripting UGS via a local socket. To enable, set the HKEY_CURRENT_USER\Software\Epic Games\UnrealGameSync\AutomationPort to a port number (eg. 30422) and restart UGS.
The generalized format of automation exchanges is:

Client:
    4 bytes: Request type (see AutomationRequestType)
    4 bytes: Request data length (X)
    X bytes: Request data

Server:
    4 bytes: Result code (see AutomationRequestResult)
    4 bytes: Response data length (X)
    X bytes: Response data

At the moment, the only request type supported, SyncProject, has a code of zero. The request data consists of two, UTF-8, null terminated strings - the stream name, and relative path to the file within in (starting with a '/' character). On success, the response will be the path to the project file on the local machine. The request will prompt the user for the workspace they would like to use for the sync (defaulting to any suitable workspace they may already have), change streams if necessary, and sync it with their current settings.

#rb none
#jira UE-65269, FORT-123041

[CL 4563331 by Ben Marsh in Dev-Build branch]