Linux: Revert CL 4523990 as this is causing the final callstack to possibly hit an unknown chunk when symbolicating
Due to x86(-64) instructions the call sites point to the next instruction. So we -1 for all call sites that is not the crash call site. This means we need
to avoid the signal handler as well to not subtract one to the place we actually crashed in. Adding to the ignore count causes us to lose this guess of -1

#jira none
#rb none
#fyi Arciel.Rekman
#lockdown Cristina.Riveron

[CL 4572786 by Brandon Schaefer in 4.21 branch]