#jira UE-69749 - Level deleted upon leaving a multi-users session.
When a session is left, the multi-user system performs some clean up to remove 'session only' transactions. If a level was created in a multi-users session and the session left while that world package was dirty, the system would revert those transactions and that process involve deleting/reloading the map. The delete action was unexpectedly recorded by the multi-user system as transaction (map deleted) because the sytem was still hooked to the editor events when this happen. Unregistring the multi-user calls back before performing the clean up fixed the issue.

When a level that was created in a session (and only exists within that session) was the active world when the user left the session, the editor would not unload it. The level could be edited, but could not be saved. Reset the world to the default level template when this situation is detected.

#rb Jamie.Dale
#fyi Jamie.Dale

[CL 4973819 by Patrick Laflamme in 4.22 branch]